requirements.txt
anyio==4.0.0
certifi==2023.7.22
click==8.1.7
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
numpy==1.26.1
pandas==2.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
